+++ src/App.js
@@ -1,14 +1,14 @@
 import { Routes, Route } from "react-router-dom";
 import Home from "./pages/HomePage/Home";
 import "./styles/partials/_global.scss";
-import MembersChart from "./pages/MembersChartPage";
+import MembersCard from "./pages/MembersCardPage";
 
 
 function App() {
   return (
     <Routes>
       <Route path="/" element={<Home />} />
-      <Route path="/members" element={<MembersChart />} />
+      <Route path="/members" element={<MembersCard />} />
     </Routes>
   );
 }
